+++ src/components/10-atoms/checkbox/story.js
@@ -22,7 +22,12 @@ storyCheckbox.add(
     const refId = text('refId', `checkbox-${createRefId()}`);
     const label = text('label', 'I agree to conditions of data protection.');
     const name = text('name', 'my-checkbox');
-    const variant = select('variant', ['square', 'checkmark']);
+    const variant = select('variant', [
+      'square',
+      'checkmark',
+      // 'inverted-square', // not officially supported yet
+      'checkmark-inverted',
+    ]);
     const checked = boolean('checked', true);
     const disabled = boolean('disabled', false);
     const errortext = boolean('error', false);
@@ -31,18 +36,24 @@ storyCheckbox.add(
 
     const wrapper = document.createElement('div');
     const template = html`
-      <axa-checkbox
-        refId="${refId}"
-        class="hover"
-        name="${name}"
-        variant="${variant}"
-        label="${label}"
-        ?disabled="${disabled}"
-        ?checked="${checked}"
-        ?required="${required}"
-        ?styled="${styled}"
-        error="${errortext ? 'Please accept our terms and conditions.' : ''}"
-      ></axa-checkbox>
+      <div
+        style="${variant && variant.includes('inverted')
+          ? `background-color: #027180; padding: 10px;`
+          : ''}"
+      >
+        <axa-checkbox
+          refId="${refId}"
+          class="hover"
+          name="${name}"
+          variant="${variant}"
+          label="${label}"
+          ?disabled="${disabled}"
+          ?checked="${checked}"
+          ?required="${required}"
+          ?styled="${styled}"
+          error="${errortext ? 'Please accept our terms and conditions.' : ''}"
+        ></axa-checkbox>
+      </div>
     `;
 
     render(template, wrapper);