+++ src/components/10-atoms/checkbox/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.0
+
+- **BREAKING CHANGE**: We do not show error messages anymore if element is disabled.
+- A new variant was added. Now you can use `checkmark-inverted` to have a better contrast on background `blue-teal`. (#1685)
+- We hardened the usage of childs as a checkbox label (internal refactoring because of bug #1699)
+
 ## 3.0.0
 
 - (This version do not have a technical breaking change.)